feat: 1. modify europarl, adding the ro-en data 2. fix a small bug of utils 3. small change for transformer (add encoder param) 4. add training for ro-en baseline model

+++ nmt/preprocess/corpus/europarl.py
@@ -11,8 +11,9 @@ def get(url, file_name, lan_1_name, lan_2_name):
     lan_2_file_path = os.path.join(europarl_news_dir, lan_2_name)
 
     utils.download(url, europarl_news_path)
-    shutil.unpack_archive(europarl_news_path, europarl_news_dir)
-    os.remove(europarl_news_path)
+    if os.path.isfile(europarl_news_path):
+        shutil.unpack_archive(europarl_news_path, europarl_news_dir)
+        os.remove(europarl_news_path)
 
     # read data
     lan_1_data = utils.read_lines(lan_1_file_path)
@@ -32,14 +33,25 @@ def de_en():
     return get(url, file_name, 'europarl-v7.de-en.de', 'europarl-v7.de-en.en')
 
 
+def ro_en():
+    file_name = 'europarl_v8_en-ro.tgz'
+    url = 'http://data.statmt.org/wmt16/translation-task/training-parallel-ep-v8.tgz'
+    return get(url, file_name, 'training-parallel-ep-v8/europarl-v8.ro-en.ro',
+               'training-parallel-ep-v8/europarl-v8.ro-en.en')
+
+
 if __name__ == '__main__':
     def show(name_1, name_2, lan_1_data, lan_2_data):
         print('\nlen of {} data: {}'.format(name_1, len(lan_1_data)))
         print('len of {} data: {}'.format(name_2, len(lan_2_data)))
 
 
-    show('de', 'en', *de_en())
-    show('fr', 'en', *fr_en())
+    show('ro', 'en', *ro_en())
+    # show('fr', 'en', *fr_en())
+    # show('fr', 'en', *fr_en())
+
+    # len of ro data: 399375
+    # len of en data: 399375
 
     # len of de data: 45913
     # len of en data: 45913
